Move fixtures to a dedicated folder and extract common variable from tests
