+++ tests/test_parser.py
@@ -1,9 +1,9 @@
 from saul import parser
-from tests import log_fixtures
+from tests.fixtures import log_fixtures
+
+log_parser = parser.LogParser()
 
 def test_parse_single_commit_log_into_list_of_changed_files():
-    log_parser = parser.LogParser()
-
     changed_files = log_parser.commit(log_fixtures.COMMIT_LOG)
 
     assert 4 == len(changed_files)
@@ -13,8 +13,6 @@ def test_parse_single_commit_log_into_list_of_changed_files():
     assert 'path/to/tests/test_order.py' == changed_files[3]
 
 def test_ignore_merge_commits():
-    log_parser = parser.LogParser()
-
     changed_files = log_parser.commit(log_fixtures.MERGE_COMMIT_LOG)
 
     assert 0 == len(changed_files)
